projects/can_profile/requirements.txt
Django==2.2
djangorestframework==3.9.1
pytz==2022.6
sqlparse==0.4.3
